Moves fetching from post page to atom-info -> add ability to fetch personas in the inventory on demand
