+++ webofneeds/won-owner-webapp/src/main/webapp/app/components/atom-info.jsx
@@ -1,8 +1,9 @@
 import React, { useState, useEffect } from "react";
 import PropTypes from "prop-types";
-import { useSelector } from "react-redux";
+import { actionCreators } from "~/app/actions/actions";
+import { useDispatch, useSelector } from "react-redux";
 import { useHistory } from "react-router-dom";
-import { get } from "../utils.js";
+import { get, generateLink } from "../utils.js";
 import WonAtomHeaderBig from "./atom-header-big.jsx";
 import WonAtomMenu from "./atom-menu.jsx";
 import WonAtomActions from "./atom-actions.jsx";
@@ -13,7 +14,8 @@ import * as processUtils from "../redux/utils/process-utils";
 import * as connectionUtils from "../redux/utils/connection-utils";
 
 import "~/style/_atom-info.scss";
-import { generateLink } from "../utils";
+import ico_loading_anim from "~/images/won-icons/ico_loading_anim.svg";
+import ico16_indicator_error from "~/images/won-icons/ico16_indicator_error.svg";
 
 export default function WonAtomInfo({
   atom,
@@ -22,13 +24,17 @@ export default function WonAtomInfo({
   initialTab = "DETAIL",
 }) {
   const history = useHistory();
+  const dispatch = useDispatch();
   const atomUri = get(atom, "uri");
   const connectionUri = get(ownedConnection, "uri");
   const processState = useSelector(generalSelectors.getProcessState);
   const storedAtoms = useSelector(generalSelectors.getAtoms);
-
+  //TODO: IMPLEMENT VISIBILITY SENSOR TO FETCH ONLY WHEN IN VIEW
   const atomLoading =
     !atom || processUtils.isAtomLoading(processState, atomUri);
+  const atomFailedToLoad =
+    atom && processUtils.hasAtomFailedToLoad(processState, atomUri);
+  const atomToLoad = !atom || processUtils.isAtomToLoad(processState, atomUri);
 
   const [visibleTab, setVisibleTab] = useState(initialTab);
   const [showAddPicker, toggleAddPicker] = useState(false);
@@ -46,6 +52,7 @@ export default function WonAtomInfo({
     },
     [atomUri, initialTab]
   );
+
   useEffect(
     () => {
       toggleActions(
@@ -54,6 +61,10 @@ export default function WonAtomInfo({
             (!!ownedConnection &&
               !connectionUtils.isConnected(ownedConnection)))
       );
+
+      if (atomUri && ((atomToLoad && !atomLoading) || !atom)) {
+        dispatch(actionCreators.atoms__fetchUnloadedAtom(atomUri));
+      }
     },
     [atomUri, connectionUri, atomLoading]
   );
@@ -73,45 +84,79 @@ export default function WonAtomInfo({
       : tabName =>
           history.replace(generateLink(history.location, { tab: tabName }));
 
-  return (
-    <won-atom-info
-      class={
-        (className ? className : "") + (atomLoading ? " won-is-loading " : "")
-      }
-    >
-      <WonAtomHeaderBig
-        atom={atom}
-        ownedConnection={ownedConnection}
-        showActions={showActions}
-        toggleActions={toggleActions}
-      />
-      {showActions ? (
-        <WonAtomActions
+  function tryReload() {
+    if (atomUri && atomFailedToLoad) {
+      dispatch(actionCreators.atoms__fetchUnloadedAtom(atomUri));
+    }
+  }
+
+  if (atomFailedToLoad) {
+    return (
+      <won-atom-info
+        class={(className ? className : "") + " won-failed-to-load "}
+      >
+        <svg className="ai__failed__icon">
+          <use xlinkHref={ico16_indicator_error} href={ico16_indicator_error} />
+        </svg>
+        <span className="ai__failed__label">
+          Failed To Load - Atom might have been deleted
+        </span>
+        <div className="ai__failed__actions">
+          <button
+            className="ai__failed__actions__button red won-button--outlined thin"
+            onClick={tryReload}
+          >
+            Try Reload
+          </button>
+        </div>
+      </won-atom-info>
+    );
+  } else if (atomLoading || atomToLoad || !atom) {
+    return (
+      <won-atom-info class={(className ? className : "") + " won-is-loading "}>
+        <svg className="ai__loading__spinner hspinner">
+          <use xlinkHref={ico_loading_anim} href={ico_loading_anim} />
+        </svg>
+        <span className="ai__loading__label">Loading...</span>
+      </won-atom-info>
+    );
+  } else {
+    return (
+      <won-atom-info class={className ? className : ""}>
+        <WonAtomHeaderBig
           atom={atom}
           ownedConnection={ownedConnection}
+          showActions={showActions}
+          toggleActions={toggleActions}
+        />
+        {showActions ? (
+          <WonAtomActions
+            atom={atom}
+            ownedConnection={ownedConnection}
+            storedAtoms={storedAtoms}
+          />
+        ) : (
+          undefined
+        )}
+        <WonAtomMenu
+          atom={atom}
+          visibleTab={visibleTab}
+          setVisibleTab={changeTab}
+          toggleAddPicker={toggleAddPicker}
+          relevantConnectionsMap={relevantConnectionsMap}
+        />
+        <WonAtomContent
+          atom={atom}
+          visibleTab={visibleTab}
+          relevantConnectionsMap={relevantConnectionsMap}
+          toggleAddPicker={toggleAddPicker}
+          showAddPicker={showAddPicker}
+          setVisibleTab={changeTab}
           storedAtoms={storedAtoms}
         />
-      ) : (
-        undefined
-      )}
-      <WonAtomMenu
-        atom={atom}
-        visibleTab={visibleTab}
-        setVisibleTab={changeTab}
-        toggleAddPicker={toggleAddPicker}
-        relevantConnectionsMap={relevantConnectionsMap}
-      />
-      <WonAtomContent
-        atom={atom}
-        visibleTab={visibleTab}
-        relevantConnectionsMap={relevantConnectionsMap}
-        toggleAddPicker={toggleAddPicker}
-        showAddPicker={showAddPicker}
-        setVisibleTab={changeTab}
-        storedAtoms={storedAtoms}
-      />
-    </won-atom-info>
-  );
+      </won-atom-info>
+    );
+  }
 }
 WonAtomInfo.propTypes = {
   atom: PropTypes.object,
