+++ webofneeds/won-owner-webapp/src/main/webapp/style/_atom-info.scss
@@ -29,4 +29,43 @@ won-atom-info {
     background: $won-secondary-color;
     color: white;
   }
+  &.won-is-loading {
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    justify-content: center;
+    padding: 4rem 0;
+    > .ai__loading__spinner.hspinner {
+      @include fixed-square(5rem);
+    }
+    > .ai__loading__label {
+      color: $won-line-gray;
+      text-align: center;
+    }
+  }
+  &.won-failed-to-load {
+    display: grid;
+    grid-template-areas: "failed_icon failed_label" "failed_actions failed_actions";
+    grid-template-rows: min-content min-content;
+    align-items: center;
+    justify-content: center;
+    padding: 4rem 0.5rem;
+    grid-gap: 0.5rem;
+    > .ai__failed__icon {
+      grid-area: failed_icon;
+      @include fixed-square(5rem);
+      --local-primary: #{$won-primary-color};
+    }
+    > .ai__failed__label {
+      grid-area: failed_label;
+      text-align: center;
+    }
+    > .ai__failed__actions {
+      grid-area: failed_actions;
+      display: flex;
+      justify-content: center;
+      > .ai__failed__actions__button {
+      }
+    }
+  }
 }
