+++ webofneeds/won-owner-webapp/src/main/webapp/style/_post.scss
@@ -28,47 +28,4 @@ main.atomcontent {
       padding: 0.5rem;
     }
   }
-
-  & > .pc__loading {
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    justify-content: center;
-    padding: 4rem 0;
-
-    > .pc__loading__spinner.hspinner {
-      @include fixed-square(5rem);
-    }
-    > .pc__loading__label {
-      color: $won-line-gray;
-      text-align: center;
-    }
-  }
-
-  & > .pc__failed {
-    display: grid;
-    grid-template-areas: "failed_icon failed_label" "failed_actions failed_actions";
-    align-items: center;
-    justify-content: center;
-    padding: 4rem 0.5rem;
-    grid-gap: 0.5rem;
-
-    > .pc__failed__icon {
-      grid-area: failed_icon;
-      @include fixed-square(5rem);
-      --local-primary: #{$won-primary-color};
-    }
-    > .pc__failed__label {
-      grid-area: failed_label;
-      text-align: center;
-    }
-    > .pc__failed__actions {
-      grid-area: failed_actions;
-      display: flex;
-      justify-content: center;
-
-      > .pc__failed__actions__button {
-      }
-    }
-  }
 }
